# output/api-contracts-updated.yaml
openapi: 3.0.0
info:
  title: FormFit Coach API
  description: |
    RESTful API for FormFit Coach platform - connecting fitness enthusiasts with expert trainers for personalized video form analysis.

    ## Authentication
    All authenticated endpoints require either:
    - Session cookie (for web app)
    - JWT token in Authorization header: `Bearer <token>`

    ## Rate Limiting
    Rate limits vary by endpoint and authentication status. Limits are returned in headers:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Requests remaining
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

    ## Versioning
    API is versioned via URL path (e.g., /api/v1/). Version v1 is current stable version.

  version: 1.0.0
  contact:
    name: FormFit Coach API Support
    email: api-support@formfit.com
servers:
  - url: https://formfit.com/api/v1
    description: Production server
  - url: https://staging.formfit.com/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: clp123abc456
        email:
          type: string
          format: email
          example: user@example.com
        name:
          type: string
          example: John Doe
        role:
          type: string
          enum: [USER, TRAINER, ADMIN]
        fitnessLevel:
          type: string
          enum: [BEGINNER, INTERMEDIATE, ADVANCED]
          nullable: true
        goals:
          type: string
          nullable: true
        emailVerified:
          type: boolean
        image:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    Trainer:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        user:
          $ref: '#/components/schemas/User'
        bio:
          type: string
          nullable: true
        specialties:
          type: array
          items:
            type: string
          example: ["Powerlifting", "Olympic Weightlifting"]
        certifications:
          type: array
          items:
            type: string
        rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
        totalReviews:
          type: integer
        isVerified:
          type: boolean
        isActive:
          type: boolean
        hourlyRate:
          type: number
          format: float
          nullable: true
        avgResponseTimeHours:
          type: number
          format: float
          nullable: true
        availableHours:
          type: object
          nullable: true
...
    Notification:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
          enum: [FEEDBACK_READY, VIDEO_ASSIGNED, UPLOAD_SUCCESS, PURCHASE_COMPLETE]
        title:
          type: string
        message:
          type: string
        link:
          type: string
          nullable: true
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
    Purchase:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        stripeSessionId:
          type: string
        amount:
          type: integer
          description: Amount in cents
        credits:
          type: integer
        status:
          type: string
          enum: [pending, completed, failed, refunded]
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
...
    PaginatedResponse:
      type: object
      properties:
        nextCursor:
          type: string
          nullable: true
        hasMore:
          type: boolean
paths:
...
  /purchases:
    get:
      tags:
        - Credits
      summary: Get purchase history
      description: Retrieve user's purchase history with pagination
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
      responses:
        '200':
          description: Purchase history
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      purchases:
                        type: array
                        items:
                          $ref: '#/components/schemas/Purchase'
                      totalSpent:
                        type: integer
                        description: Total amount spent in cents
        '401':
          description: Unauthorized
  /notifications:
    get:
      tags:
        - Notifications
      summary: List notifications
      description: Get user's notifications with pagination
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of notifications
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      notifications:
                        type: array
                        items:
                          $ref: '#/components/schemas/Notification'
                      unreadCount:
                        type: integer
        '401':
          description: Unauthorized
...
  /notifications/poll:
    get:
      tags:
        - Notifications
      summary: Poll for new notifications
      description: Lightweight endpoint for polling notification count
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: since
          in: query
          schema:
            type: string
            format: date-time
          description: Only count notifications after this timestamp
      responses:
        '200':
          description: Notification poll response
          content:
            application/json:
              schema:
                type: object
                properties:
                  unreadCount:
                    type: integer
                  hasNew:
                    type: boolean
                  latestNotification:
                    $ref: '#/components/schemas/Notification'
                    nullable: true
        '401':
          description: Unauthorized
  /user/profile:
    patch:
      tags:
        - Users
      summary: Update user profile
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 100
                fitnessLevel:
                  type: string
                  enum: [BEGINNER, INTERMEDIATE, ADVANCED]
                goals:
                  type: string
                  maxLength: 500
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  updated:
                    type: array
                    items:
                      type: string
        '400':
          description: Validation error
        '401':
          description: Unauthorized
...
  /user/password:
    patch:
      tags:
        - Users
      summary: Change password
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currentPassword
                - newPassword
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Invalid current password or password requirements not met
        '401':
          description: Unauthorized
        '429':
          description: Rate limit exceeded (max 5 per hour)
  /trainer/profile:
    patch:
      tags:
        - Trainers
      summary: Update trainer profile
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bio:
                  type: string
                  maxLength: 500
                specialties:
                  type: array
                  items:
                    type: string
                hourlyRate:
                  type: number
                  minimum: 0
                availableHours:
                  type: object
                maxDailyReviews:
                  type: integer
                  minimum: 1
                  maximum: 50
      responses:
        '200':
          description: Trainer profile updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  trainer:
                    $ref: '#/components/schemas/Trainer'
                  updated:
                    type: array
                    items:
                      type: string
        '401':
          description: Unauthorized
        '403':
          description: Not a trainer
...